+++ engine/src/main/java/dev/gamekit/core/Window.java
@@ -2,12 +2,13 @@ package dev.gamekit.core;
 
 import dev.gamekit.settings.Resolution;
 import dev.gamekit.settings.Settings;
-import dev.gamekit.utils.Position;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import javax.swing.*;
 import java.awt.*;
+import java.awt.event.FocusAdapter;
+import java.awt.event.FocusEvent;
 import java.awt.image.BufferStrategy;
 import java.awt.image.BufferedImage;
 
@@ -19,22 +20,27 @@ public final class Window {
 
   private final int displayWidth;
   private final int displayHeight;
+  private final int centerX;
+  private final int centerY;
   private final JFrame frame;
+  private final Canvas canvas;
   private final BufferStrategy bufferStrategy;
   private final BufferedImage displayBuffer;
   private final Graphics2D displayGraphics;
   private final BufferedImage uiBuffer;
   private final Graphics2D uiGraphics;
-  private final Info info;
-  private final Position[] renderPositions;
 
   Window() {
-    double displayScaleRatio;
     Settings settings = Application.getInstance().getSettings();
+    GraphicsEnvironment ge = GraphicsEnvironment.getLocalGraphicsEnvironment();
+    GraphicsConfiguration gc = ge.getDefaultScreenDevice().getDefaultConfiguration();
 
     displayWidth = settings.resolution.width();
     displayHeight = settings.resolution.height();
+    centerX = displayWidth / 2;
+    centerY = displayHeight / 2;
     frame = new JFrame(settings.title);
+    canvas = new Canvas();
 
     if (settings.fullscreen) {
       Dimension d = new Dimension(Resolution.NATIVE.width(), Resolution.NATIVE.height());
@@ -44,23 +50,28 @@ public final class Window {
       frame.setResizable(false);
       frame.setUndecorated(true);
 
-      GraphicsEnvironment.getLocalGraphicsEnvironment().getDefaultScreenDevice().setFullScreenWindow(frame);
-
-      displayScaleRatio = Math.min(
-        (double) Resolution.NATIVE.width() / displayWidth,
-        (double) Resolution.NATIVE.height() / displayHeight
-      );
+      ge.getDefaultScreenDevice().setFullScreenWindow(frame);
     } else {
       Dimension d = new Dimension(displayWidth, displayHeight);
+
       frame.setMinimumSize(d);
       frame.setPreferredSize(d);
       frame.setResizable(false);
       frame.setUndecorated(settings.undecorated);
-      displayScaleRatio = 1;
     }
 
-    GraphicsEnvironment ge = GraphicsEnvironment.getLocalGraphicsEnvironment();
-    GraphicsConfiguration gc = ge.getDefaultScreenDevice().getDefaultConfiguration();
+    Dimension d = new Dimension(displayWidth, displayHeight);
+    canvas.setSize(d);
+    canvas.setMinimumSize(d);
+    canvas.setMaximumSize(d);
+    canvas.setPreferredSize(d);
+    canvas.setFocusTraversalKeysEnabled(false);
+    canvas.addFocusListener(new FocusAdapter() {
+      @Override
+      public void focusLost(FocusEvent e) {
+        canvas.requestFocusInWindow();
+      }
+    });
 
     displayBuffer = gc.createCompatibleImage(displayWidth, displayHeight, Transparency.TRANSLUCENT);
     displayGraphics = displayBuffer.createGraphics();
@@ -79,32 +90,20 @@ public final class Window {
     settings.dithering.apply(uiGraphics);
 
     frame.setIgnoreRepaint(true);
-    frame.getContentPane().setBackground(Color.BLACK);
+    frame.setLayout(new GridBagLayout());
+    frame.setBackground(Color.BLACK);
+    frame.getContentPane().add(canvas);
     frame.setLocationRelativeTo(null);
     frame.pack();
 
-    frame.createBufferStrategy(2);
-    bufferStrategy = frame.getBufferStrategy();
-
-    info = new Info(
-      frame.getWidth(), frame.getHeight(),
-      displayWidth, displayHeight,
-      displayWidth / 2, displayHeight / 2,
-      displayScaleRatio, 1.0 / displayScaleRatio
-    );
-
-    int scaledWidth = (int) (displayWidth * displayScaleRatio);
-    int scaledHeight = (int) (displayHeight * displayScaleRatio);
-    Position topLeft = new Position(
-      (int) (0.5 * (frame.getWidth() - scaledWidth)),
-      (int) (0.5 * (frame.getHeight() - scaledHeight))
-    );
-    Position bottomRight = new Position(topLeft.x + scaledWidth, topLeft.y + scaledHeight);
-
-    renderPositions = new Position[]{ topLeft, bottomRight };
+    canvas.createBufferStrategy(2);
+    bufferStrategy = canvas.getBufferStrategy();
 
     LOGGER.debug("Created window");
-    LOGGER.debug(info);
+    LOGGER.debug(
+      "displayWidth={}, displayHeight={}, centerX={}, centerY={}",
+      displayWidth, displayHeight, centerX, centerY
+    );
 
     Window.instance = this;
   }
@@ -114,14 +113,34 @@ public final class Window {
     return instance;
   }
 
-  /** Returns the {@link Info} object associated with the current instance of {@link Window} */
-  public Info getInfo() {
-    return info;
+  /** Returns the display width of the {@link Window} */
+  public int getDisplayWidth() {
+    return displayWidth;
+  }
+
+  /** Returns the display height of the {@link Window} */
+  public int getDisplayHeight() {
+    return displayHeight;
+  }
+
+  /** Returns the x-coordinate of the center point of the {@link Window} */
+  public int getCenterX() {
+    return centerX;
+  }
+
+  /** Returns the y-coordinate of the center point of the {@link Window} */
+  public int getCenterY() {
+    return centerY;
   }
 
   /** Returns the visible {@link JFrame} of the {@link Window} */
   JFrame getFrame() {
     return frame;
+  }
+
+  /** Returns the visible {@link Canvas} of the {@link Window} */
+  Canvas getCanvas() {
+    return canvas;
   }
 
   /** Returns the {@link Graphics2D} of the display buffer to which the scene is rendered */
@@ -137,43 +156,17 @@ public final class Window {
   /** Makes the associated {@link JFrame} visible */
   void show() {
     frame.setVisible(true);
+    canvas.requestFocus();
   }
 
   /** Updates the {@link JFrame} buffer strategy with the display and UI buffers */
   void update() {
-    Position tl = renderPositions[0];
-    Position br = renderPositions[1];
-
     do {
       Graphics2D bufferGraphics = (Graphics2D) bufferStrategy.getDrawGraphics();
-      bufferGraphics.drawImage(displayBuffer, tl.x, tl.y, br.x, br.y, 0, 0, displayWidth, displayHeight, null);
-      bufferGraphics.drawImage(uiBuffer, tl.x, tl.y, br.x, br.y, 0, 0, displayWidth, displayHeight, null);
+      bufferGraphics.drawImage(displayBuffer, 0, 0, displayWidth, displayHeight, null);
+      bufferGraphics.drawImage(uiBuffer, 0, 0, displayWidth, displayHeight, null);
       bufferStrategy.show();
       bufferGraphics.dispose();
     } while (bufferStrategy.contentsLost());
-  }
-
-  /** {@link Window.Info} holds the read-only parameters of the current {@link Window instance} */
-  public record Info(
-    int frameWidth,
-    int frameHeight,
-    int displayWidth,
-    int displayHeight,
-    int displayCenterX,
-    int displayCenterY,
-    double displayScaleRatio,
-    double inverseDisplayScaleRatio
-  ) {
-    @Override
-    public String toString() {
-      return String.format(
-        "%s[frame=(%dx%d), display=(%dx%d), center=[%d,%d], scaleRatio=%.2f, inverseScaleRatio=%.2f]",
-        getClass().getName(),
-        frameWidth, frameHeight,
-        displayWidth, displayHeight,
-        displayCenterX, displayCenterY,
-        displayScaleRatio, inverseDisplayScaleRatio
-      );
-    }
   }
 }